refactor: Button to new architecture
