+++ src/components/Button/index.ts
@@ -1,0 +1,1 @@
+export { Button } from './Button'
